requirements.txt
APScheduler==3.6.3
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
MarkupSafe==1.1.1
numpy==1.19.4
pandas==1.1.4
python-aqi==0.6.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
requests==2.24.0
scipy==1.5.4
six==1.15.0
tzlocal==2.1
urllib3==1.25.11
Werkzeug==1.0.1
xgboost==1.2.1
